Refined the code for the feedback bar and more
